[Refactoring] Split each routing handlers for testing
