+++ routers/api/users/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const express = require('express');
+const shouldUserLogin = require('../../../middlewares/shouldUserLogin');
+
+const router = express.Router();
+
+router.use(shouldUserLogin);
+
+router.get('/me', require('./me'));
+
+module.exports = router;
