+++ routers/users/default_image.js
@@ -1,13 +1,8 @@
 'use strict';
 
 const path = require('path');
-const express = require('express');
 
-const router = express.Router();
-
-router.get('/default_image(.:format)', function (req, res) {
+module.exports = function (req, res) {
   const format = req.param.format || 'png';
   res.sendFile(path.resolve(__dirname, `../assets/img-user_default.${format}`));
-});
-
-module.exports = router;
+};
